requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
anyio==3.2.1
    # via httpcore
backcall==0.2.0
    # via ipython
certifi==2021.5.30
    # via httpx
click==7.1.2
    # via typer
coloredlogs==15.0.1
    # via sxm_player (pyproject.toml)
decorator==5.0.9
    # via ipython
fake-useragent==0.1.11
    # via sxm
greenlet==1.1.0
    # via sqlalchemy
h11==0.12.0
    # via httpcore
httpcore==0.13.6
    # via httpx
httpx==0.18.2
    # via
    #   sxm
    #   sxm_player (pyproject.toml)
humanfriendly==9.2
    # via coloredlogs
idna==3.2
    # via
    #   anyio
    #   rfc3986
ipython==7.25.0
    # via sxm_player (pyproject.toml)
ipython-genutils==0.2.0
    # via traitlets
jedi==0.18.0
    # via ipython
matplotlib-inline==0.1.2
    # via ipython
parso==0.8.2
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.19
    # via ipython
psutil==5.8.0
    # via sxm_player (pyproject.toml)
ptyprocess==0.7.0
    # via pexpect
pydantic==1.8.2
    # via sxm
pygments==2.9.0
    # via ipython
python-dotenv==0.18.0
    # via sxm
pyyaml==5.4.1
    # via sxm_player (pyproject.toml)
rfc3986[idna2008]==1.5.0
    # via httpx
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==1.4.21
    # via sxm_player (pyproject.toml)
sxm==0.2.3
    # via sxm_player (pyproject.toml)
tenacity==8.0.1
    # via sxm
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
typer==0.3.2
    # via
    #   sxm
    #   sxm_player (pyproject.toml)
typing-extensions==3.10.0.0
    # via pydantic
ua-parser==0.10.0
    # via sxm
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
